Rewrite metrics-batch in Perl
Since the intermediate steps of acumulating one .yml file per directory
does is not needed anymore to generate the final CSV file, they were removed.

But then now we are missing some some caching scheme to avoid the having
to reprocess all directories. This should be fixed by #11 in a near
future.