Prevent adding the same commit twice
Although there will be no commits with the same id for a given project,
this allows to reprocess a project against the same output database. The
source tree will still be processed, but the database output will just
ignore it.

It would be nice if the output object could say "I already have this
job" and the runner would just skip it. This way we would be able to
*update* a DB with the latest commits from a given project without the
need to reprocess all commits that were already processed before.