Install non-Perl files (like templates, images, etc) under a shared directory #31
File::Share + File::ShareDir + File::ShareDir::Install provides all
needed tools to keep static and non-code files under specific shared
directories to it.

The File::Share::dist_dir() looks for a local shared directory, if it
notice that is running on a development environment. Otherwise it look
in the system.

Assuming we have a file under "share/file.txt" we can find it doing:

  package Analizo::Sample::ShareDir;
  use Analizo
  use File::Share ':all';
  my $share = dist_dir('Analizo');
  open FILE, '<', "$share/file.txt";
  ...
  return 1;

With the code above doesn't matter if we are running on development
environment or in production with code installed via package for
example.

Another small changes made in this commit:

* Distribuition name in Makefile.PL: analizo -> Analizo
* Run tests using prove when building Debian package
* Update MANIFEST file
* Skip MYMETA.yml when update MANIFEST