Removing Analizo::Output::DOT in order to use Graph::Writer::Dot
The Graph::Writer::Dot is on the CPAN and it is a mature piece of
software, it do almost the same job of the Analizo::Output::DOT, then
in my opinion is much better to use the Graph::Writer::Dot instead of
maintain the same code in the Analizo side.

But Graph::Writer::Dot doesn't have a important feature implemented in
the Analizo::Output::DOT to cluster the graph in subgraphs. That way I
implemented this feature in Graph::Writer::Dot and sent it to the
upstream author.

While the upstream author doesn't review my code I'm adding here (in the
Analizo side) the file "lib/Graph/Writer/Dot.pm" that implements the
Graph::Writer::Dot.