Fixing running of a single acceptance test
Cucumber wasn't running a single acceptance test under t/features/metrics/
because it's impossible to load step definitions from a path out of the
directory where the *.feature file is loaded.

The strategy to solve this problem is to count the number of steps loaded
after calling Cucumber load method. If the Hash of steps is empty, it calls
the method load_steps through the t/features/step_definitions folder.

Now it is possible to load steps when calling sigle tests from any folder.

Signed-off-by: Luan Guimarães <livreluan@gmail.com>
Signed-off-by: Izabela Cardoso <cizabelacristina@gmail.com>