improving and arranging ACC and Graph Tests
This change was necessary because the tests were broken due to the high coupling between each ACC unit test.
The Model.pm tests were breaking due to the modifications in the calls of the graphs created by it, in addition,
it was used to create tests for the new generated graph.
Also added a new feature test for ACC with deep inheritance.

Co-authored-by: Allan Nobre <allanjefrey9@hotmail.com>
Co-authored-by: Lucas Costa <lucascostaa73@gmail.com>