Revert "Add extractor for Python"
This reverts commit c9486ea2b3dd0ed10bf45e960dfd9a8ed5686949.

Reverting on the main (stable) branch to prepare a bugfix release.

This commit is preserved on the dev (unstable) branch which
contains work-in-progress features, and the commit
c9486ea2b3dd0ed10bf45e960dfd9a8ed5686949 depends on Pyan which is
not yet available on Debian or other distributions where Analizo is
already packaged.